requirements.txt
django-rest-framework==0.1.0
markdown2
whitenoise